renamed gameLogic package
